Split processing and default preprocesisng params into separate files
