+++ src/super_gradients/training/datasets/samplers/class_balanced_sampler.py
@@ -127,7 +127,13 @@ class ClassBalancedSampler(WeightedRandomSampler):
         oversample_aggressiveness: float = 0.5,
         num_samples: Optional[int] = None,
         generator=None,
+        shuffle: bool = True,  # noqa
     ) -> None:
+        """
+
+        :param shuffle: This parameter is not used, exists for compatibility purposes
+               since dataloaders.get() forwards this argument to samples
+        """
         """
         Wrap WeightedRandomSampler with weights that are computed from the class frequencies of the dataset.
         """
